Refactor out withChangeTracking
