+++ packages/suggestion/src/slate-diff/internal/transforms/transformDiffTexts.ts
@@ -1,5 +1,4 @@
 import {
-  addRangeMarks,
   isText,
   PlateEditor,
   TDescendant,
@@ -7,340 +6,13 @@ import {
   TText,
   Value,
 } from '@udecode/plate-common';
-import isEqual from 'lodash/isEqual.js';
-import uniqWith from 'lodash/uniqWith.js';
-import {
-  BaseEditor,
-  createEditor,
-  Editor,
-  InsertTextOperation,
-  MergeNodeOperation,
-  Node,
-  Path,
-  Point,
-  PointRef,
-  Range,
-  RangeRef,
-  RemoveTextOperation,
-  SetNodeOperation,
-  SplitNodeOperation,
-  withoutNormalizing,
-} from 'slate';
+import { createEditor, Path, withoutNormalizing } from 'slate';
 
 import { DiffToSuggestionsOptions } from '../../slateDiff';
 import { dmp } from '../utils/dmp';
 import { getProperties } from '../utils/get-properties';
 import { InlineNodeCharMap } from '../utils/inline-node-char-map';
-
-interface NodesEditor extends BaseEditor {
-  propsChanges: {
-    rangeRef: RangeRef;
-    properties: Record<string, any>;
-    newProperties: Record<string, any>;
-  }[];
-
-  insertedTexts: {
-    rangeRef: RangeRef;
-    node: TText;
-  }[];
-
-  removedTexts: {
-    pointRef: PointRef;
-    node: TText;
-  }[];
-
-  commitDiffs: () => void;
-}
-
-const objectWithoutUndefined = (obj: Record<string, any>) => {
-  const newObj: Record<string, any> = {};
-
-  Object.keys(obj).forEach((key) => {
-    if (obj[key] !== undefined) {
-      newObj[key] = obj[key];
-    }
-  });
-
-  return newObj;
-};
-
-const flattenPropsChanges = (editor: NodesEditor) => {
-  // The set of points at which some `propChanges` range starts or ends
-  const unsortedRangePoints = editor.propsChanges.flatMap(({ rangeRef }) => {
-    const range = rangeRef.current;
-    if (!range) return [];
-    return [range.anchor, range.focus];
-  });
-
-  const rangePoints = uniqWith(
-    unsortedRangePoints.sort(Point.compare),
-    Point.equals
-  );
-  if (rangePoints.length < 2) return [];
-
-  /**
-   * A continuous set of non-overlapping ranges spanning the first and last
-   * `rangePoints`.
-   */
-  const flatRanges = Array.from({ length: rangePoints.length - 1 })
-    .fill(null)
-    .map((_, i) => ({
-      anchor: rangePoints[i],
-      focus: rangePoints[i + 1],
-    }));
-
-  const flatUpdates = flatRanges.map((flatRange) => {
-    // The set of `propChanges` that intersect with `flatRange`
-    const intersectingUpdates = editor.propsChanges.filter(({ rangeRef }) => {
-      const range = rangeRef.current;
-      if (!range) return false;
-      const intersection = Range.intersection(range, flatRange);
-      if (!intersection) return false;
-      return Range.isExpanded(intersection);
-    });
-
-    if (intersectingUpdates.length === 0) return null;
-
-    // Get the props of the range before and after the updates
-    const initialProps = objectWithoutUndefined(
-      intersectingUpdates[0].properties
-    );
-
-    const finalProps = objectWithoutUndefined(
-      intersectingUpdates.at(-1)!.newProperties
-    );
-
-    if (isEqual(initialProps, finalProps)) return null;
-
-    const properties = {} as Record<string, any>;
-    const newProperties = {} as Record<string, any>;
-
-    for (const key of Object.keys(finalProps)) {
-      if (!isEqual(initialProps[key], finalProps[key])) {
-        properties[key] = initialProps[key];
-        newProperties[key] = finalProps[key];
-      }
-    }
-
-    for (const key of Object.keys(initialProps)) {
-      if (finalProps[key] === undefined) {
-        properties[key] = initialProps[key];
-        newProperties[key] = undefined;
-      }
-    }
-
-    return {
-      range: flatRange,
-      properties,
-      newProperties,
-    };
-  });
-
-  editor.propsChanges.forEach(({ rangeRef }) => {
-    rangeRef.unref();
-  });
-
-  return flatUpdates.filter(Boolean) as Exclude<
-    (typeof flatUpdates)[number],
-    null
-  >[];
-};
-
-const withNodesEditor = (
-  editor: NodesEditor,
-  {
-    getInsertProps,
-    getRemoveProps,
-    getUpdateProps,
-  }: Required<DiffToSuggestionsOptions>
-) => {
-  const { apply } = editor;
-
-  editor.propsChanges = [];
-  editor.insertedTexts = [];
-  editor.removedTexts = [];
-
-  let recordingOperations = true;
-
-  const applyInsertText = (op: InsertTextOperation) => {
-    const node = Node.get(editor, op.path) as TText;
-
-    apply(op);
-
-    const startPoint = { path: op.path, offset: op.offset };
-    const endPoint = { path: op.path, offset: op.offset + op.text.length };
-    const range = { anchor: startPoint, focus: endPoint };
-    const rangeRef = Editor.rangeRef(editor, range);
-
-    editor.insertedTexts.push({
-      rangeRef,
-      node: {
-        ...node,
-        text: op.text,
-      },
-    });
-  };
-
-  const applyRemoveText = (op: RemoveTextOperation) => {
-    const node = Node.get(editor, op.path) as TText;
-
-    apply(op);
-
-    const point = { path: op.path, offset: op.offset };
-    const pointRef = Editor.pointRef(editor, point, {
-      affinity: 'backward',
-    });
-
-    editor.removedTexts.push({
-      pointRef,
-      node: {
-        ...node,
-        text: op.text,
-      },
-    });
-  };
-
-  const applyMergeNode = (op: MergeNodeOperation) => {
-    const oldNode = Node.get(editor, op.path) as TText;
-    const properties = Node.extractProps(oldNode);
-
-    const prevNodePath = Path.previous(op.path);
-    const prevNode = Node.get(editor, prevNodePath) as TText;
-    const newProperties = Node.extractProps(prevNode);
-
-    apply(op);
-
-    const startPoint = { path: prevNodePath, offset: prevNode.text.length };
-    const endPoint = Editor.end(editor, prevNodePath);
-    const range = { anchor: startPoint, focus: endPoint };
-    const rangeRef = Editor.rangeRef(editor, range);
-
-    editor.propsChanges.push({
-      rangeRef,
-      properties,
-      newProperties,
-    });
-  };
-
-  const applySplitNode = (op: SplitNodeOperation) => {
-    const oldNode = Node.get(editor, op.path) as TText;
-    const properties = Node.extractProps(oldNode);
-    const newProperties = op.properties;
-
-    apply(op);
-
-    const newNodePath = Path.next(op.path);
-    const newNodeRange = Editor.range(editor, newNodePath);
-    const rangeRef = Editor.rangeRef(editor, newNodeRange);
-
-    editor.propsChanges.push({
-      rangeRef,
-      properties,
-      newProperties,
-    });
-  };
-
-  const applySetNode = (op: SetNodeOperation) => {
-    apply(op);
-
-    const range = Editor.range(editor, op.path);
-    const rangeRef = Editor.rangeRef(editor, range);
-
-    editor.propsChanges.push({
-      rangeRef,
-      properties: op.properties,
-      newProperties: op.newProperties,
-    });
-  };
-
-  editor.apply = (op) => {
-    if (!recordingOperations) {
-      return apply(op);
-    }
-
-    recordingOperations = false;
-
-    switch (op.type) {
-      case 'insert_text': {
-        applyInsertText(op);
-        break;
-      }
-
-      case 'remove_text': {
-        applyRemoveText(op);
-        break;
-      }
-
-      case 'merge_node': {
-        applyMergeNode(op);
-        break;
-      }
-
-      case 'split_node': {
-        applySplitNode(op);
-        break;
-      }
-
-      case 'set_node': {
-        applySetNode(op);
-        break;
-      }
-
-      default: {
-        apply(op);
-      }
-    }
-
-    recordingOperations = true;
-  };
-
-  editor.commitDiffs = () => {
-    recordingOperations = false;
-
-    editor.removedTexts.forEach(({ pointRef, node }) => {
-      const point = pointRef.current;
-
-      if (point) {
-        editor.insertNode(
-          {
-            ...node,
-            ...getRemoveProps(node),
-          },
-          { at: point }
-        );
-      }
-
-      pointRef.unref();
-    });
-
-    editor.insertedTexts.forEach(({ rangeRef, node }) => {
-      const range = rangeRef.current;
-
-      if (range) {
-        addRangeMarks(editor as any, getInsertProps(node), { at: range });
-      }
-
-      rangeRef.unref();
-    });
-
-    // Reverse the array to prevent path changes
-    const flatUpdates = flattenPropsChanges(editor).reverse();
-
-    flatUpdates.forEach(({ range, properties, newProperties }) => {
-      const node = Node.get(editor, range.anchor.path) as TText;
-
-      addRangeMarks(
-        editor as any,
-        getUpdateProps(node, properties, newProperties),
-        { at: range }
-      );
-    });
-
-    recordingOperations = true;
-  };
-
-  return editor;
-};
+import { withChangeTracking } from '../utils/with-change-tracking';
 
 // Main function to transform an array of text nodes into another array of text nodes
 export function transformDiffTexts<
@@ -370,7 +42,7 @@ export function transformDiffTexts<
   const texts = nodes.map((n) => inlineNodeCharMap.nodeToText(n));
   const nextTexts = nextNodes.map((n) => inlineNodeCharMap.nodeToText(n));
 
-  const nodesEditor = withNodesEditor(createEditor() as any, options);
+  const nodesEditor = withChangeTracking(createEditor(), options);
   nodesEditor.children = [{ children: texts }];
 
   withoutNormalizing(nodesEditor, () => {
@@ -396,7 +68,7 @@ export function transformDiffTexts<
       nodesEditor.apply(op);
     }
 
-    nodesEditor.commitDiffs();
+    nodesEditor.commitChangesToDiffs();
   });
 
   const diffTexts: TText[] = (nodesEditor.children[0] as any).children;
